moved Loading Dialog class to administartion package
